chore: rename control file
