+++ package/lib/src/controls/echo_text.dart
